requirements.txt
celery==5.6.3
Django==6.0.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-rest-knox==5.0.4
djangorestframework==3.17.1
gunicorn==25.2.0
pillow==12.1.1
psycopg2-binary==2.9.11
